refactor(phase-2w): extract OpenProjectWorkPackageCustomFieldService
Continues openproject_client.py god-class decomposition (ADR-002).
Two work-package CF write helpers move to
``OpenProjectWorkPackageCustomFieldService`` exposed as
``self.wp_cf``.

Methods moved
-------------
- bulk_set_wp_custom_field_values (was on
  OpenProjectCustomFieldService — that service is about CF schema
  ops; per-work-package value writes belong with the work-package
  writes)
- set_wp_last_update_date_by_keys (was on OpenProjectClient — used
  by the change-detection pipeline)

Security analysis
-----------------
bulk_set_wp_custom_field_values
- Inputs (wp_id, cf_id, value) are coerced via int()/str() in a
  Python list comprehension before reaching Ruby; embedded inside a
  single-quoted ``<<-'J2O_DATA'`` heredoc parsed by ``JSON.parse``,
  so contents are data, not code. SAFE.
- FIX: dropped trailing ``.to_json`` on the final Ruby expression.
  ``execute_query_to_json_file`` already wraps the script tail with
  ``.as_json`` Ruby-side; the explicit ``.to_json`` would have
  double-encoded and Python would receive a JSON string, not a
  parsed dict.
- FIX: moved the int()/str() coercion inside a try/except so a
  malformed row returns the documented
  ``{success: False, updated, failed, error}`` envelope rather
  than raising KeyError/TypeError/ValueError past the caller.

set_wp_last_update_date_by_keys
- ``project_id`` was bare-interpolated as ``proj_id = {project_id}``.
  FIX: ``int(project_id)`` is now performed inside the existing
  try/except so coercion failures fall through the documented
  ``{updated: 0, examined: 0, error}`` envelope. The integer is
  embedded as a Ruby integer literal.
- ``date_str`` was interpolated raw inside Ruby single quotes
  (``target_date = '{date_str}'``). FIX: now passed through
  ``escape_ruby_single_quoted`` so quote/backslash injection cannot
  break out of the literal.
- ``jira_keys`` previously round-tripped through
  ``json.dumps(json.dumps(list(keys)))`` to embed a JSON-string-of-
  a-JSON-array, then Ruby parsed the outer string literal back to
  the inner JSON. FIX: the keys list is now embedded inside a
  single-quoted ``<<-'J2O_KEYS'`` heredoc parsed by ``JSON.parse``,
  matching the pattern used elsewhere in the codebase. Same data,
  much clearer contract, no double encoding.
- The script returns a Ruby hash literal (no trailing ``.to_json``
  in the original; left unchanged).

Numbers
-------
- openproject_client.py: 2562 -> 2505 LOC
- openproject_custom_field_service.py: 600 -> 520 LOC
  (bulk_set_wp_custom_field_values moved out)
- openproject_work_package_custom_field_service.py: 0 -> 241 LOC

Verification
------------
- pytest tests/unit: 953 passed
- mypy src/: clean
- ruff check / ruff format: clean
